+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added Warden
   * Increased sonic boom range and darkness range
+* Added Beta Creepers
 * Increased mounting chance
 * Slightly slowed down miner mobs
 * Fixed Angry creeper integration with ITR